Reorganizing the instance 'class' sources
