+++ src/defaults.js
@@ -3,8 +3,7 @@ const DEFAULTS = {
   appSecret: process.env.CRAFT_APP_SECRET,
   destroyOnExit: true,
   httpApiUrl: process.env.CRAFT_HTTP_API_URL || 'https://api.craft.ai/v1',
-  wsApiUrl: process.env.CRAFT_WS_API_URL || 'wss://api.craft.ai/v1',
-  hubApiUrl: process.env.CRAFT_HUB_API_URL || 'https://hub.craft.ai/v1'
+  wsApiUrl: process.env.CRAFT_WS_API_URL || 'wss://api.craft.ai/v1'
 };
 
 export default DEFAULTS;
